requirements.txt
fastapi
uvicorn[standard]
pillow
sentence-transformers
faiss-cpu
pytesseract
python-multipart
tqdm
